queries/Flask-Sqagger-EmployeeCRUDApp/requirements.txt

flask
flask-restx
flask-sqlalchemy
python-dotenv
